Separate current tab test from keyboard shortcuts
