+++ src/ui/popup.html
@@ -15,11 +15,11 @@
 </head>
 <body>
 <template id="template-current-tab-button">
-  <button type="submit" class="dropdown-item" name="action" value="export-current-tab" data-format="custom-format" data-custom-format-slot="">Current tab </button>
+  <button type="submit" class="dropdown-item" name="action" value="export-current-tab" data-format="custom-format" data-custom-format-slot="" id="">Current tab </button>
 </template>
 <template id="template-export-tabs-button">
-  <button type="submit" class="dropdown-item" name="action" value="export-tabs" data-scope="all" data-format="custom-format" data-custom-format-slot="">All tabs </button>
-  <button type="submit" class="dropdown-item" name="action" value="export-tabs" data-scope="highlighted" data-format="custom-format" data-custom-format-slot="">Selected tabs </button>
+  <button type="submit" class="dropdown-item" name="action" value="export-tabs" data-scope="all" data-format="custom-format" data-custom-format-slot="" id="">All tabs </button>
+  <button type="submit" class="dropdown-item" name="action" value="export-tabs" data-scope="highlighted" data-format="custom-format" data-custom-format-slot="" id="">Selected tabs </button>
 </template>
 <div class="dropdown is-active">
   <div class="dropdown-menu is-relative p-0" id="dropdown-menu" role="menu">
